requirements.txt
fastapi
sqlmodel
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
passlib[bcrypt]
python-multipart
httpx